+++ Song.java
@@ -1,9 +1,9 @@
 public class Song {
     private String title;
     private String artist;
-    private int playCount;
+    private long playCount;
 
-public Song(String title, String artist, int playCount) {
+public Song(String title, String artist, long playCount) {
     this.title = title;
     this.artist = artist;
     this.playCount = playCount;
@@ -17,7 +17,7 @@ public String getArtist() {
     return artist;
 }
 
-public int getPlayCount() {
+public long getPlayCount() {
     return playCount;
 }
